fix(sdk): single file host.js
