+++ typegraph/core/src/utils/postprocess/wasmedge_rt.rs
@@ -2,6 +2,7 @@
 // SPDX-License-Identifier: MPL-2.0
 
 use crate::utils::fs_host;
+use crate::wit::metatype::typegraph::host::read_file;
 use common::{
     archive::encode_bytes_to_base_64,
     typegraph::{
@@ -27,7 +28,7 @@ impl PostProcessor for WasmedgeProcessor {
                 };
 
                 let wasi_path = fs_host::make_absolute(&PathBuf::from(path))?;
-                let bytes = crate::wit::read_file(&wasi_path.display().to_string())?;
+                let bytes = read_file(&wasi_path.display().to_string())?;
                 mat_data.wasm = encode_bytes_to_base_64(bytes).map_err(|e| e.to_string())?;
 
                 mat.data = map_from_object(mat_data).map_err(|e| e.to_string())?;
