+++ typegraph/core/wit/typegraph.wit
@@ -499,14 +499,19 @@ interface utils {
     oauth2-with-custom-profiler: func(service-name: string, scopes: string, profiler: type-id) -> result<string, error>
 }
 
+interface host {
+    printf: func(s: string)
+    expand-glob: func(root: string, exclude: list<string>) -> result<list<string>, string>
+    read-file: func(path: string) -> result<list<u8>, string>
+    write-file: func(path: string, data: list<u8>) -> result<_, string>
+    get-cwd: func() -> result<string, string>
+}
+
+
 world typegraph {
     export core
     export runtimes
     export utils
     export aws
-    import print: func(s: string)
-    import expand-glob: func(root: string, exclude: list<string>) -> result<list<string>, string>
-    import read-file: func(path: string) -> result<list<u8>, string>
-    import write-file: func(path: string, data: list<u8>) -> result<_, string>
-    import get-cwd: func() -> result<string, string>
+    import host
 }
