+++ typegraph/node/sdk/src/host/host.ts
@@ -2,6 +2,8 @@
 // SPDX-License-Identifier: MPL-2.0
 import * as fs from "node:fs";
 import * as path from "node:path";
+import * as process from "node:process";
+
 function listAllFilesHelper(
   root: string,
   list: Array<string>,
@@ -24,12 +26,44 @@ function listAllFilesHelper(
   }
 }
 
-export default function (root: string, exclude: Array<string>): Array<string> {
+export function expandGlob(
+  root: string,
+  exclude: Array<string>,
+): Array<string> {
   try {
     const ret = [] as Array<string>;
     // console.log("[host] received args", root, exclude);
     listAllFilesHelper(root, ret, exclude);
     return ret;
+  } catch (err) {
+    throw (err instanceof Error ? err.message : err);
+  }
+}
+
+export function print(msg: string) {
+  console.log("[host]", msg);
+}
+
+export function getCwd(): string {
+  try {
+    return process.cwd();
+  } catch (err) {
+    throw (err instanceof Error ? err.message : err);
+  }
+}
+
+export function readFile(path: string): Uint8Array {
+  try {
+    const buffer = fs.readFileSync(path, null);
+    return new Uint8Array(buffer);
+  } catch (err) {
+    throw (err instanceof Error ? err.message : err);
+  }
+}
+
+export function writeFile(path: string, data: Uint8Array): void {
+  try {
+    void fs.writeFileSync(path, data);
   } catch (err) {
     throw (err instanceof Error ? err.message : err);
   }
