--- typegraph/node/sdk/src/host/print.ts
@@ -1,6 +1,0 @@
-// Copyright Metatype OÜ, licensed under the Mozilla Public License Version 2.0.
-// SPDX-License-Identifier: MPL-2.0
-
-export default function (msg: string) {
-  console.log("[host]", msg);
-}
