+++ whiz.yaml
@@ -54,7 +54,7 @@ typegraph:
     wasm-tools component new target/wasm/wasm32-unknown-unknown/debug/typegraph_core.wasm -o $WASM_FILE
 
     rm -rf typegraph/node/sdk/src/gen typegraph/node/sdk/dist
-    jco transpile $WASM_FILE -o typegraph/node/sdk/src/gen --map print=../host/print.js read-file=../host/read_file.js write-file=../host/write_file.js expand-glob=../host/expand_glob.js get-cwd=../host/get_cwd.js
+    jco transpile $WASM_FILE -o typegraph/node/sdk/src/gen --map metatype:typegraph/host=../host/host.js
 
     cd typegraph/node
     pnpm run sdk-build
